add warning for unsupported go-whatsapp
Added a warning as this package is not currently being maintained. This should stop users from attempting to use it while multi-device rolls out and experiencing errors, or developing their own application to use this library, only to have it become unusable in a few weeks.

Any edits to the message are welcome.